+++ dsmr_mqtt/management/commands/dsmr_mqtt.py
@@ -2,7 +2,6 @@ import logging
 
 from django.core.management.base import BaseCommand
 from django.utils.translation import gettext as _
-from django.conf import settings
 
 from dsmr_backend.mixins import InfiniteManagementCommandMixin, StopInfiniteRun
 from dsmr_mqtt.models.settings.broker import MQTTBrokerSettings
@@ -15,13 +14,13 @@ logger = logging.getLogger('commands')
 class Command(InfiniteManagementCommandMixin, BaseCommand):
     help = _('Dedicated MQTT client for publishing messages to the broker.')
     name = __name__  # Required for PID file.
-    sleep_time = settings.DSMRREADER_MQTT_SLEEP
 
     # Global during this process' lifetime.
     mqtt_client = None
 
     def initialize(self):
         """ Set up persistent MQTT client. """
+        self.sleep_time = MQTTBrokerSettings.get_solo().process_sleep
         self.mqtt_client = dsmr_mqtt.services.broker.initialize()
 
     def shutdown(self):
@@ -31,10 +30,8 @@ class Command(InfiniteManagementCommandMixin, BaseCommand):
 
     def run(self, **options):
         """ InfiniteManagementCommandMixin listens to handle() and calls run() in a loop. """
-        broker_settings = MQTTBrokerSettings.get_solo()
-
         # Check on each run. In case MQTT was either disabled, enabled or settings were changed.
-        if broker_settings.restart_required:
+        if MQTTBrokerSettings.get_solo().restart_required:
             MQTTBrokerSettings.objects.update(restart_required=False)
             logger.warning('MQTT | --- Detected settings change, requiring process restart, stopping...')
             raise StopInfiniteRun()