+++ dsmrreader/config/dsmrreader.py
@@ -24,9 +24,9 @@ DSMRREADER_MODULE_STATS_GENERATOR = 'dsmr_stats.services.run'
 DSMRREADER_MODULE_MINDERGAS_EXPORT = 'dsmr_mindergas.services.run'
 
 # Sleep durations for infinity processes. Update these in your own config if you wish to alter them.
-DSMRREADER_BACKEND_SLEEP = 1
-DSMRREADER_DATALOGGER_SLEEP = 0.5
-DSMRREADER_MQTT_SLEEP = 1
+DSMRREADER_BACKEND_SLEEP = 1  # Will be REMOVED in v3.x+
+DSMRREADER_DATALOGGER_SLEEP = 0.5  # Will be REMOVED in v3.x+
+DSMRREADER_MQTT_SLEEP = 1  # Will be REMOVED in v3.x+
 
 DSMRREADER_DROPBOX_MAX_FILE_MODIFICATION_TIME = 60 * 60 * 24 * 7
 DSMRREADER_DROPBOX_SYNC_INTERVAL = 1  # Only check for changes once per hour.
@@ -35,10 +35,10 @@ DSMRREADER_DROPBOX_ERROR_INTERVAL = 12  # Skip new files for 12 hours when insuf
 DSMRREADER_COMPACT_MAX = 1024  # Max telegrams to compact in a single run.
 
 # Whether telegrams are logged, in base64 format. Only required for debugging.
-DSMRREADER_LOG_TELEGRAMS = False
+DSMRREADER_LOG_TELEGRAMS = False  # Will be REMOVED in v3.x+
 
-# Whether the backend process (and datalogger) reconnects to the DB after each run.
-DSMRREADER_RECONNECT_DATABASE = True
+# When processes should reconnect to the DB, to ensure the connection is still there.
+DSMRREADER_MAX_DATABASE_CONNECTION_SESSION_IN_SECONDS = 30 * 60
 
 # Maximum interval allowed since the latest reading, before ringing any alarms.
 DSMRREADER_STATUS_READING_OFFSET_MINUTES = 60
@@ -54,8 +54,5 @@ DSMRREADER_RETENTION_MAX_CLEANUP_HOURS_PER_RUN = 24
 
 # Plugins.
 DSMRREADER_PLUGINS = []
-
-# Whether to override (disable) capabilities.
-DSMRREADER_DISABLED_CAPABILITIES = []
 
 DSMRREADER_BUIENRADAR_API_URL = 'https://data.buienradar.nl/2.0/feed/json'